requirements.txt
alembic==1.15.2
catboost==1.2.8
click==8.1.8
email_validator==2.2.0
fastapi==0.115.12
pandas==2.2.3
pydantic==2.11.3
python-dotenv==1.1.0
python-jose==3.4.0
scikit-learn==1.6.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
uvicorn==0.34.2
passlib==1.7.4
python-multipart==0.0.20
pyodbc==5.2.0
bcrypt==4.0.1
passlib==1.7.4
pydantic[email]
jose
psycopg2-binary==2.9.10